entry: move flatpak building into actions
